+++ save.txt
@@ -1,6 +1,2 @@
-T | 1 | hello
-T | 0 | burn
-T | 1 | byebye
-D | 0 | return book  | 2019-10-15
-E | 0 | party  | 2025-08-27T18:00 | 2025-08-27T20:00
-E | 0 | party  | 2025-08-27T18:00 | 2025-08-27T22:00
+T | 0 | money
+D | 0 | return books  | 2019-09-22
